chore: split facility service and data access tests
